% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-1-QRS.R
\name{QRS}
\alias{QRS}
\title{Quantile Rank Score Test based on gene-specific delta}
\usage{
QRS(x, z, y, subject, tau = 0.5)
}
\arguments{
\item{x}{the covariate for the nuisance parameters, and x does not contain the intercept.}

\item{z}{the covariate for the parameter(s) of interest.}

\item{y}{the response variable.}

\item{subject}{indicates the ID or number of arrays/subjects.}

\item{tau}{the quantile level.}
}
\value{
Returns a \code{list} object that contains:

  QRSi : p-value of the quantile rank score test which ignores the dependence entirely, such that delta = tau^2

  QRS0 : p-value of the quantile rank score test, where delta is estimated by using the residuals obtained under H0

  QRS1 : p-value of the quantile rank score test, where delta is estimated by using the residuals obtained under H1

  sn   : the quantile rank score

  Qni, Qn0, Qn1: the variance-covariance of sn for methods QRSi, QRS0 and QRS1, respectively

  Tni, Tn0, Tn1: the test statistic values for methods QRSi, QRS0 and QRS1, respectively

  beta: the quantile estimate for z
}
\description{
Quantile Rank Score Test based on gene-specific delta
}
