% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-2-TTS.R
\name{TTS}
\alias{TTS}
\title{A tail-based test statistics (TTS).}
\usage{
TTS(n, j, data, tau)
}
\arguments{
\item{n}{xxx}

\item{j}{xxx}

\item{data}{xxx}

\item{tau}{xxx}
}
\value{
Returns a \code{list} object that the model solution.
}
\description{
A tail-based test statistics (TTS).
}
