% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-0-ttsDE.R
\docType{package}
\name{ttsDE-package}
\alias{ttsDE-package}
\title{ttsDE: A tail-based test statistics (TTS) for differential expression detection}
\description{
TTS is a tail-based test to identify diffentially expressed genes between
 two groups. TTS is a quantile-based model that takes expression values at
 exon level, and adjusts for covariates.
}
\section{Function}{

ttsDE
}

