% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-3-ttsDE.R
\name{ttsDE}
\alias{ttsDE}
\title{A tail-based test statistics (TTS) for differential expression detection.}
\usage{
ttsDE(z, d, c, gene, tau, method = c("TTS", "QRS", "QS", "lme"), n_core = 1)
}
\arguments{
\item{z}{A matrix of gene expressions.}

\item{d}{A vector of group allocation.}

\item{c}{Covariates.}

\item{gene}{A character vector of gene names, corresponding to each row of z.}

\item{tau}{Pre-specified tau.}

\item{method}{Methods included.}

\item{n_core}{Number of cores in parallel computing.}
}
\value{
Returns a \code{list} object that the model solution.
}
\description{
A tail-based test statistics (TTS) for differential expression detection.
}
